integration_2023_w10b

See merge request oai/openairinterface5g!2002

* !1954 Fix for warnings and errors from CLANG compiling
* !1978 Change default of "nokrnmod" option to "On"
* !1981 F1AP: minor fixes and preparation for F1 work
* !1983 Improve clang-format
* !1992 fix nrqtscope build warnings
* !1994 reworking response to MSG3 with DCCH/DTCH
* !1996 dead-code-f1-removal
* !1968 small changes for multi usrp support
* !1991 rfsimulator: configurable wait timeout, use in CI to avoid retransmissions
* !1999 Remove Dedicated Serving Cell Configuration from CI config file for 2x2 rfsim test

+++ openair2/LAYER2/NR_MAC_UE/mac_defs.h
@@ -505,7 +505,7 @@ typedef struct prach_association_pattern {
 typedef struct ssb_info {
   bool transmitted; // True if the SSB index is transmitted according to the SSB positions map configuration
   prach_occasion_info_t *mapped_ro[MAX_NB_RO_PER_SSB_IN_ASSOCIATION_PATTERN]; // List of mapped RACH Occasions to this SSB index
-  uint16_t nb_mapped_ro; // Total number of mapped ROs to this SSB index
+  uint32_t nb_mapped_ro; // Total number of mapped ROs to this SSB index
 } ssb_info_t;
 
 // List of all the possible SSBs and their details
